major code refactoring
