+++ frontend/src/locales/en.json
@@ -41,46 +41,7 @@
       "title-placeholder": "Event Title"
     }
   },
-  "pageMode": {
-    "organizer": "organizer",
-    "participant": "participant"
-  },
   "pages": {
-    "about": {
-      "community": {
-        "p1": "But the development itself wouldn't be possible without the great help of our friends and the Matchy community, consistently giving us valuable feedback on how to improve and motivating us to stay on track.",
-        "p2": "We really hope that this project brings you joy and helps you make fun memories and lasting relationships.",
-        "title": "🧑‍🤝‍🧑 The Community"
-      },
-      "devs": {
-        "daniel": {
-          "p1": "Daniel is the rapid prototyper in the team with the most experience in the software industry. He developed essential features reliably with a high quality.",
-          "title": "The Rapid Prototyper."
-        },
-        "description": "The development team is currently composed of the following main contributors:",
-        "marco": {
-          "p1": "Marco took the responsibility of designing the application architecture and implementing the backend almost entirely by himself (Postgres database, Postgres CRON jobs, API via Supabase, Lambda functions, RPCs, user input validation ...).",
-          "p2": "He also created many useful diagrams that helped us understand the real time processes that this application is built on, implemented huge chunks of the frontend and improved the overall performance and security of this application greatly.",
-          "title": "The Architect."
-        },
-        "stefan": {
-          "p1": "Stefan is the most knowledgeable frontend developer in the team. He reviewed every single of our pull requests and helped us get a better understanding of advanced Vue.js concepts.",
-          "p2": "He decided which technologies to use for the frontend, set up all necessary dependencies and always found the most elegant and efficient solutions to UI / UX problems.",
-          "p3": "He is the teams frontend Sensei because he was really patient with newcomers and taught them everything they needed to know about Vue.js.",
-          "title": "The Web Developer."
-        },
-        "title": "The Devs",
-        "yahya": {
-          "p1": "This app was initially his idea and his second attempt based on a previous iteration called {0}.",
-          "p2": "After defining the necessary features for this iteration and designing the mockups, Yahya gathered all his friends to team up and kick off.",
-          "p3": "He kept the bigger picture at all times by contributing to all aspects of the application and managed the project effectively.",
-          "title": "The Kickstarter."
-        }
-      },
-      "github-page": "Github page",
-      "github-plug": "Check the out this project's {0} to see the source-code and find out who we are, what we're all about and what we are aiming to achieve.",
-      "title": "🌱 About Us"
-    },
     "dashboard": {
       "ongoing": {
         "round": "Round",
