+++ frontend/src/stores/auth.ts
@@ -2,36 +2,33 @@ import { acceptHMRUpdate, defineStore } from "pinia";
 import { supabase } from "@/services/supabase";
 import type { Provider, User } from "@supabase/supabase-js";
 import { useProfileService, type Profile } from "@/services/profileService";
-import { storageRef } from "./utils/localstorage";
+import { storageRef } from "./localstorage/localstorage";
 
 export const useAuthStore = defineStore("user", () => {
-  // user state (only update with setter)
+  // User
   const user = ref<User | null>(supabase.auth.user());
   const isLoggedIn = computed(() => user.value !== null);
 
   const redirect = storageRef("redirect");
 
-  // user setter
   async function setUserStore(newUser: User | null) {
     console.log("Updating user state", newUser);
     user.value = newUser;
     if (newUser) {
-      const fetchedProfile = await useProfileService().readProfile(newUser);
+      const fetchedProfile = await useProfileService().fetchProfile(newUser);
       setProfileStore(fetchedProfile);
     }
   }
 
-  // user profile state (only update with setter)
+  // Profile
   const profile = ref<Profile>({});
   const isRegistered = computed(() => !!profile.value.fullName);
 
-  // profile setter
   function setProfileStore(newProfile: Profile) {
     console.log("Updating profile state", newProfile);
     profile.value = newProfile;
   }
 
-  // on authState change, update everything
   supabase.auth.onAuthStateChange((event, session) => {
     console.log("Auth session changed", event, session);
     setUserStore(session?.user ?? null);
@@ -41,20 +38,15 @@ export const useAuthStore = defineStore("user", () => {
     const redirectURL =
       new URL(import.meta.env.BASE_URL, window.location.origin) + "callback";
     redirect.value = "/edit-profile";
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { error, user, session } = await supabase.auth.signUp(
-      {
-        email,
-        password,
-      },
-      {
-        redirectTo: redirectURL,
-      }
+      { email, password },
+      { redirectTo: redirectURL }
     );
 
     if (error) throw error;
   }
 
-  // LOGIN
   async function login(email: string, password: string) {
     const { error } = await supabase.auth.signIn({ email, password });
     if (error) throw error;
@@ -64,19 +56,16 @@ export const useAuthStore = defineStore("user", () => {
     redirect.value = afterLoginRedirect;
     const redirectTo =
       new URL(import.meta.env.BASE_URL, window.location.origin) + "callback";
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const { user, session, error } = await supabase.auth.signIn(
-      {
-        provider: provider,
-      },
+      { provider: provider },
       { redirectTo }
     );
     if (error) throw error;
   }
 
-  // LOGOUT
   function logout() {
-    // TODO: Catch error
-    supabase.auth.signOut();
+    supabase.auth.signOut(); // throws error
   }
 
   async function deleteAccount() {
@@ -85,8 +74,8 @@ export const useAuthStore = defineStore("user", () => {
       errorToast(error);
       throw error;
     }
-    // we have to call this otherwise the user is never updated
-    // but it throws an error because the user does not exist anymore 😂
+    // Weird quirk: we have to call this - otherwise the user is never updated.
+    // But it will throw an error because the user does not exist anymore (because we deleted it).
     logout();
   }
