--- frontend/src/styles/global.css
@@ -1,11 +1,0 @@
-/* CHROME BROWSER FIX: no background-color change for form fields after autofill */
-input:-webkit-autofill,
-input:-webkit-autofill:hover,
-input:-webkit-autofill:focus,
-input:-webkit-autofill:active {
-  transition: background-color 5000s ease-in-out 0s;
-}
-
-html {
-  background-color: #f6f6f6;
-}
